+++ R/tt_export.R
@@ -58,8 +58,10 @@ collapse_path <- function(paths) {
 }
 
 collapse_values <- function(colvals) {
-    if(!is.list(colvals) || all(vapply(colvals, length, 1L) == 1))
+    if(!is.list(colvals)) ## || all(vapply(colvals, length, 1L) == 1))
         return(colvals)
+    else if(all(vapply(colvals, length, 1L) == 1))
+        return(unlist(colvals))
     vapply(colvals, paste, "", collapse = .collapse_char)
 }
 
@@ -126,7 +128,7 @@ path_enriched_df <- function(tt, path_fun = collapse_path, value_fun = collapse_
 #' }
 export_as_txt <- function(tt, file = NULL, paginate = FALSE, ..., page_break = "\\s\\n") {
 
-    colwidths <- propose_column_widths(tt)
+    colwidths <- propose_column_widths(matrix_form(tt, indent_rownames = TRUE))
 
     if(paginate) {
         tbls <- paginate_table(tt, ...)
@@ -168,8 +170,9 @@ export_as_txt <- function(tt, file = NULL, paginate = FALSE, ..., page_break = "
 #' ft <- tt_to_flextable(tbl)
 #' ft
 
-tt_to_flextable <- function(tt, paginate = FALSE, lpp = NULL, ..., colwidths = propose_column_widths(tt),
-                            total_width = 5) {
+tt_to_flextable <- function(tt, paginate = FALSE, lpp = NULL, ...,
+                            colwidths = propose_column_widths(matrix_form(tt, indent_rownames = TRUE)),
+                            total_width = 10) {
     if(!requireNamespace("flextable") || !requireNamespace("officer")) {
         stop("this function requires the flextable and officer packages. Please install them if you wish to use it")
     }
@@ -178,7 +181,6 @@ tt_to_flextable <- function(tt, paginate = FALSE, lpp = NULL, ..., colwidths = p
     if(paginate) {
         if(is.null(lpp))
             stop("lpp must be specified when calling tt_to_flextable with paginate=TRUE")
-        colwidths <- propose_column_widths(tt)
         tabs <- paginate_table(tt, lpp = lpp, ...)
         return(lapply(tabs, tt_to_flextable, paginate=FALSE, total_width = total_width, colwidths = colwidths))
     }
@@ -281,7 +283,7 @@ export_as_pdf <- function(tt,
     grid.newpage()
     pushViewport(plotViewport(margins = margins, gp = gp_plot))
 
-    colwidths <- propose_column_widths(tt)
+    colwidths <- propose_column_widths(matrix_form(tt, indent_rownames = TRUE))
     tbls <- if (paginate) {
 
         if (is.null(lpp)) {
@@ -295,7 +297,7 @@ export_as_pdf <- function(tt,
         list(tt)
     }
 
-    stbls <- lapply(lapply(tbls, toString, widths = colwidths), function(xi) substr(xi, 1, nchar(xi) - nchar("\n")))
+    stbls <- lapply(lapply(tbls, toString, widths = colwidths, linesep = "-"), function(xi) substr(xi, 1, nchar(xi) - nchar("\n")))
 
     gtbls <- lapply(stbls, function(txt) {
         textGrob(