+++ tests/testthat/test-formatting.R
@@ -19,11 +19,32 @@ test_that("sprintf_format works correctly", {
 })
 
 
-test_that("formats supported", {
+
+test_that("table_shell works", {
+
+    tbl <- rtable(c("A", "B"),
+                  rrow("Hiya", rcell(c(2, .2), format = "xx (xx.x%)"), rcell(c(.1, .2)), format = "xx.x - xx.x"),
+                  rrow("bye", 5.2345, 17.2),
+                  format = "xx.xx")
 
 
-    forms <- list_rcell_format_labels()
+    tblsh <- rtable(c("A", "B"),
+                  rrow("Hiya", "xx (xx.x%)", "xx.x - xx.x"),
+                  rrow("bye", "xx.xx", "xx.xx"))
 
-    res <- sapply(forms, function(vc) all(sapply(vc, is_rcell_format)))
-    expect_true(all(res))
+    expect_identical(toString(tblsh),
+                     paste0(capture_output(table_shell(tbl)), "\n"))
+
+    tbl2 <-  rtable(c("A", "B"),
+                  rrow("Hiya", rcell(c(2, .2), format = function(x,...) paste0(x)), rcell(c(.1, .2)), format = "xx.x - xx.x"),
+                  rrow("bye", 5.2345, 17.2),
+                  format = "xx.xx")
+
+    tbl2sh <- rtable(c("A", "B"),
+                    rrow("Hiya", "<fnc>", "xx.x - xx.x"),
+                    rrow("bye", "xx.xx", "xx.xx"))
+
+    expect_identical(toString(tbl2sh),
+                     paste0(capture_output(table_shell(tbl2)), "\n"))
+
 })